+++ onprem/ingest/__init__.py
@@ -1,2 +1,2 @@
-from .base import Ingester, load_single_document, load_documents, process_folder, chunk_documents
+from .base import load_single_document, load_documents, process_folder, chunk_documents
 from .helpers import extract_files